androidDagger directory renamed to daggerCore
