+++ feature/backup/src/main/java/com/phelat/tedu/backup/view/WebDavSetupFragment.kt
@@ -9,7 +9,7 @@ import androidx.core.widget.addTextChangedListener
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.observe
-import com.phelat.tedu.androiddagger.inject
+import com.phelat.tedu.daggercore.inject
 import com.phelat.tedu.backup.R
 import com.phelat.tedu.backup.di.component.BackupComponent
 import com.phelat.tedu.backup.entity.WebDavCredentials
