+++ library/daggercore/src/main/AndroidManifest.xml
@@ -1,3 +1,3 @@
 <manifest
     xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.phelat.tedu.androiddagger" />
+    package="com.phelat.tedu.daggercore" />
